refactor(types): move PeerId from lib-storage to lib-types
Move the PeerId type to lib-types (canonical types crate) and re-export
from lib-storage for backward compatibility.

Changes:
- Create lib-types/src/peer.rs with PeerId enum
- Add PeerId to lib-types exports
- lib-storage/src/dht/transport.rs now re-exports from lib-types
- Maintains all existing PeerId functionality

Backward compatibility:
- Code importing from lib_storage::dht::PeerId continues to work
- Code can now also import from lib_types::PeerId

Closes #1649 (TYPES-7)
