+++ lib-types/src/lib.rs
@@ -5,6 +5,7 @@
 
 pub mod primitives;
 pub mod node_id;
+pub mod peer;
 pub mod dht;
 pub mod chunk;
 pub mod errors;
@@ -13,6 +14,7 @@ pub mod errors;
 pub use primitives::{Address, Amount, BlockHash, BlockHeight, Bps, TokenId, TxHash};
 
 pub use node_id::NodeId;
+pub use peer::PeerId;
 pub use dht::*;
 pub use chunk::*;
 pub use errors::*;
